Move 'compile_and_publish_diagnostics' from 'diagnostic_ext' to 'main'.
